+++ src/modules/Area.js
@@ -123,7 +123,7 @@ export default class Area {
    */
   scroll = (directions, multiplier) => {
     scrollElement(this._node, directions, multiplier)
-    this.PubSub.publish('Area:scroll', { directions })
+    this.PubSub.publish('Area:scroll', { data: { directions, multiplier } })
   }
 
   //////////////////////////////////////////////////////////////////////////////////////