+++ src/modules/Interaction.js
@@ -19,6 +19,8 @@ export default class Interaction {
   constructor({ areaElement, DS }) {
     this._areaElement = areaElement
     this.DS = DS
+    this.DS.subscribe('PointerStore:updated', this.update)
+    this.DS.subscribe('Area:scroll', this.update)
   }
 
   init = () => {
@@ -33,6 +35,8 @@ export default class Interaction {
   start = (event) => {
     if (event.type === 'touchstart') event.preventDefault() // Call preventDefault() to prevent double click issue, see https://github.com/ThibaultJanBeyer/DragSelect/pull/29 & https://developer.mozilla.org/vi/docs/Web/API/Touch_events/Supporting_both_TouchEvent_and_MouseEvent
     if (/** @type {*} */ (event).button === 2) return // right-clicks
+
+    this.isInteracting = true
     this.DS.publish('Interaction:start', { event })
 
     document.addEventListener('mouseup', this.reset)
@@ -40,6 +44,7 @@ export default class Interaction {
   }
 
   stop = () => {
+    this.isInteracting = false
     this._areaElement.removeEventListener('mousedown', this.start)
     this._areaElement.removeEventListener('touchstart', this.start, {
       // @ts-ignore
@@ -49,9 +54,15 @@ export default class Interaction {
     document.removeEventListener('touchend', this.reset)
   }
 
+  update = ({ event, data }) => {
+    if (this.isInteracting)
+      this.DS.publish('Interaction:update', { event, data })
+  }
+
   reset = (event) => {
     this.stop()
     this.init()
+    // debounce, make sure that the end event is put at the end of the event loop
     this.DS.publish('Interaction:end', { event })
   }
 }